+++ demo/src/app/layout/layout.component.ts
@@ -1,9 +1,8 @@
-import { EmptyState } from './../../../../src/empty-state/empty-state';
-
 import {
     Component,
     OnInit,
-    ViewEncapsulation
+    ViewEncapsulation,
+    ChangeDetectionStrategy
 } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { ActionDropdown } from './../../../../src/action-dropdown';
@@ -14,6 +13,8 @@ import { Segment } from './../../../../src/segment/segment';
 import { Fab } from './../../../../src/fabs/fab';
 import { Badge } from './../../../../src/badges/badge';
 import { Chip } from './../../../../src/chip/chip';
+import { EmptyState } from './../../../../src/empty-state/empty-state';
+import { ProfileDropdown } from './../../../../src/profile-dropdown/profile-dropdown';
 
 @Component({
     encapsulation: ViewEncapsulation.None,
@@ -21,7 +22,8 @@ import { Chip } from './../../../../src/chip/chip';
     templateUrl: 'layout.component.html',
     styleUrls: [
         './layout.component.scss'
-    ]
+    ],
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 
 export class LayoutComponent implements OnInit {
@@ -46,6 +48,38 @@ export class LayoutComponent implements OnInit {
                     icon: 'filter'
                 })
             ]
+        }
+    );
+    public profile: ProfileDropdown = new ProfileDropdown(
+        {
+            title: 'Your Profile',
+            img: '../assets/img/jonsnow.png',
+            defaultImg: '../assets/img/profile.png',
+            dropdown: [
+                new Action({
+                    label: 'Your profile',
+                    callback: () => {
+                        console.log('goToProfile');
+                    },
+                    icon: 'account'
+                }),
+                new Action({
+                    label: 'Settings',
+                    callback: () => {
+                        console.log('goToSettings');
+                    },
+                }),
+                new Action({
+                    label: 'Web Page',
+                    link: 'http://alvarosaburido.com'
+                })
+            ],
+            lastAction: new Action({
+                label: 'Logout',
+                callback: () => {
+                    console.log('logout');
+                }
+            })
         }
     );
     public header: Header = new Header(