+++ gulpfile.js
@@ -77,7 +77,7 @@ gulp.task('rollup:fesm', function () {
 
       // Bundle's entry point
       // See https://github.com/rollup/rollup/wiki/JavaScript-API#entry
-      entry: `${buildFolder}/index.js`,
+      input: `${buildFolder}/index.js`,
 
       // Allow mixing of hypothetical and actual files. "Actual" files can be files
       // accessed by Rollup or produced by plugins further down the chain.
@@ -110,7 +110,7 @@ gulp.task('rollup:umd', function () {
 
       // Bundle's entry point
       // See https://github.com/rollup/rollup/wiki/JavaScript-API#entry
-      entry: `${buildFolder}/index.js`,
+      input: `${buildFolder}/index.js`,
 
       // Allow mixing of hypothetical and actual files. "Actual" files can be files
       // accessed by Rollup or produced by plugins further down the chain.
@@ -131,12 +131,12 @@ gulp.task('rollup:umd', function () {
 
       // Export mode to use
       // See https://github.com/rollup/rollup/wiki/JavaScript-API#exports
-      exports: 'named',
+      output: 'named',
 
       // The name to use for the module for UMD/IIFE bundles
       // (required for bundles with exports)
       // See https://github.com/rollup/rollup/wiki/JavaScript-API#modulename
-      moduleName: 'ngx-canaima',
+      name: 'ngx-canaima',
 
       // See https://github.com/rollup/rollup/wiki/JavaScript-API#globals
       globals: {