+++ src/action.ts
@@ -1,36 +1,46 @@
 export class Action {
     public label?: string;
+    public active?: boolean;
     public callback?: any;
     public icon?: string;
     public iconActive?: string;
     public value?: boolean;
     public size?: string;
     public prefix?: string;
+    public activeClass?: string;
     public customClass?: string;
     public color?: string;
     public outline?: boolean;
     public rounded?: boolean;
     public circle?: boolean;
+    public anchor?: boolean;
+    public link?: string;
     constructor(
         options: {
             label?: string;
+            active?: boolean;
             callback?: any;
             icon?: string;
             iconActive?: string;
             value?: boolean;
             size?: string;
             prefix?: string;
+            activeClass?: string;
             customClass?: string;
             color?: string;
             outline?: boolean;
             rounded?: boolean;
             circle?: boolean;
+            anchor?: boolean;
+            link?: string;
         }
     ) {
         this.label = options.label;
+        this.active = options.active || false;
         this.icon = options.icon;
         this.iconActive = options.iconActive;
         this.prefix = options.prefix || 'mdi';
+        this.activeClass = options.activeClass;
         this.customClass = options.customClass;
         this.callback = options.callback;
         this.size = options.size;
@@ -39,21 +49,7 @@ export class Action {
         this.outline = options.outline;
         this.rounded = options.rounded;
         this.circle = options.circle;
+        this.anchor = options.anchor;
+        this.link = options.link;
     }
 }
-
-/*
-label = null,
-        callback = () => {
-            console.log('Simple action');
-        },
-        icon = null,
-        iconActive= null,
-        value= false,
-        customClass = null,
-        prefix = 'mdi',
-        size = 'md',
-        color = 'link',
-        outline?: boolean,
-        rounded?: boolean
-*/