+++ src/cards/cards.module.ts
@@ -1,3 +1,5 @@
+import { ToggleModule } from './../toggle/toggle.module';
+import { CardToggleComponent } from './card-toggle/card-toggle.component';
 import { MiscModule } from './../misc/misc.module';
 
 
@@ -13,15 +15,14 @@ import { CardIllustrationComponent } from './card-illustration/card-illustration
 import { CardNumberComponent } from './card-number/card-number.component';
 import { FabsModule } from '../fabs/fabs.module';
 import { ButtonsModule } from '../buttons/buttons.module';
-import { CountoModule } from 'angular2-counto';
 
 @NgModule({
   imports: [
     ButtonsModule,
+    ToggleModule,
     FabsModule,
     CommonModule,
     BsDropdownModule,
-    CountoModule,
     MiscModule
   ],
   exports: [
@@ -32,7 +33,8 @@ import { CountoModule } from 'angular2-counto';
     CardDataComponent,
     CardProfileComponent,
     CardIllustrationComponent,
-    CardNumberComponent
+    CardNumberComponent,
+    CardToggleComponent
   ],
   declarations: [
     CardComponent,
@@ -41,7 +43,8 @@ import { CountoModule } from 'angular2-counto';
     CardDataComponent,
     CardProfileComponent,
     CardIllustrationComponent,
-    CardNumberComponent
+    CardNumberComponent,
+    CardToggleComponent
   ],
   providers: []
 })