+++ src/index.ts
@@ -1,13 +1,8 @@
-import { ContentPlaceholderModule } from './content-placeholder/content-placeholder.module';
-import { MiscModule } from './misc/misc.module';
-
-
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CardsModule } from './cards/cards.module';
 import { ListsModule } from './lists/lists.module';
 import { GridListsModule } from './grid-list/grid-list.module';
-
 import { DynamicFormsModule } from './forms/dynamic-forms/dynamic-forms.module';
 import { ToolbarModule } from './toolbar/toolbar.module';
 import { HeaderModule } from './header/header.module';
@@ -24,6 +19,10 @@ import { LoadingModule } from
 './loading/loading.module';
 import { PreviewModule } from './preview/preview.module';
 import { MediaUploaderModule } from './media-uploader/media-uploader.module';
+import { DataTableModule } from './data-table/data-table.module';
+import { ProfileDropdownModule } from './profile-dropdown/profile-dropdown.module';
+import { ContentPlaceholderModule } from './content-placeholder/content-placeholder.module';
+import { MiscModule } from './misc/misc.module';
 
 export * from './action';
 export * from './action-dropdown';
@@ -44,6 +43,10 @@ export * from './search-bar';
 export * from './loading';
 export * from './empty-state';
 export * from './misc';
+export * from './profile-dropdown';
+export * from './validation';
+export * from './toggle';
+export * from './data-table';
 
 @NgModule({
   imports: [
@@ -67,10 +70,11 @@ export * from './misc';
     EmptyStateModule,
     MediaUploaderModule,
     MiscModule,
-    ContentPlaceholderModule
+    ContentPlaceholderModule,
+    ProfileDropdownModule,
+    DataTableModule
   ],
   declarations: [
-    
   ],
   exports: [
     CardsModule,
@@ -92,7 +96,9 @@ export * from './misc';
     EmptyStateModule,
     MediaUploaderModule,
     MiscModule,
-    ContentPlaceholderModule
+    ContentPlaceholderModule,
+    ProfileDropdownModule,
+    DataTableModule
   ]
 })
 export class NgxCanaimaModule {