+++ src/media-uploader/media-uploader.component.ts
@@ -24,44 +24,36 @@ declare var API_URL: string;
     templateUrl: 'media-uploader.component.html'
 })
 export class MediaUploaderComponent implements OnInit, OnChanges, AfterViewInit {
-    @Input('image') public image;
+    @Input('media') public media;
     @Input('customClass') public customClass;
     @Input('prefix') public prefix = 'mdi';
     @Input('icon') public icon = 'image';
     @Input('multiple') public multiple = false;
-    @Input('config') public config?: FileUploaderOptions = {
-      url: 'http://localhost:5678/api/medias/upload',
-      itemAlias: 'asmedia'
-    };
+    /* @Input('config') public config?: FileUploaderOptions = {
+      url: `${API_URL}/medias`,
+      itemAlias: 'body'
+    }; */
     @Input('uploader') public uploader: FileUploader;
     @Output() public uploaderChange = new EventEmitter<FileUploader>();
     @Output() public onQueue = new EventEmitter<any>();
-    
     public hasBaseDropZoneOver = false;
 
-    constructor(
-      public mediaUploaderService: MediaUploaderService,
-    ) {
-      this.mediaUploaderService.config(
-        this.config
-      );
+    constructor() {
     }
     public ngOnInit() {
       this.uploader = this.mediaUploaderService.uploader;
       setTimeout(() => {
-        if (this.uploader === undefined || this.uploader) {
-          this.uploader = new FileUploader(this.config);
-          this.uploader.onCompleteItem = (fileItem) => {
-            this.uploader = new FileUploader(this.config);
-          }
+        console.log(' Init ', this.uploader);
+        /* if (this.uploader === undefined || this.uploader) {
+          /* this.uploader = new FileUploader(this.config); 
           this.uploader.onAfterAddingFile = (fileItem) => {
             this.onQueue.emit(fileItem);
             fileItem.withCredentials = false;
           };
           this.mediaUploaderService.uploader = this.uploader;
         }
-        this.uploaderChange.emit(this.uploader);
-      }, 500);
+        this.uploaderChange.emit(this.uploader); */
+      });
     }
     public ngAfterViewInit() {
       /* this.uploader = new FileUploader({
@@ -72,12 +64,31 @@ export class MediaUploaderComponent implements OnInit, OnChanges, AfterViewInit 
 
     }
     public ngOnChanges(changes: SimpleChanges) {
-      console.log(changes)
-      if (changes.uploader.previousValue === undefined) {
-        this.uploaderChange.emit(this.uploader);
-      }
+      console.log(' Changes ', changes);
+      /* console.log(changes)
+      if (changes.deuploader.previousValue === undefined) {
+        this.deuploaderChange.emit(this.uploader);
+      } */
     }
     public fileOverBase(e: any): void {
       this.hasBaseDropZoneOver = e;
+    }
+    /**
+     * checkFileType
+     */
+    public checkFileType(file) {
+      if (file) {
+        let type = file.type.split('/')[0];
+        return type;
+      }
+    }
+    /**
+     * checkMimeType
+     */
+    public checkMimeType(file) {
+      if (file) {
+        let mime = file.type.split('/')[1];
+        return mime;
+      }
     }
 }