+++ src/misc/index.ts
@@ -1,2 +1,4 @@
 export * from './format-case.service';
 export * from './number-si-prefix.pipe';
+export * from './spacer/spacer.component';
+export * from './keys.pipe';