+++ src/misc/misc.module.ts
@@ -1,14 +1,24 @@
+
 import { NgModule } from '@angular/core';
 import { NumberSiPrefixPipe } from './number-si-prefix.pipe';
 import { FormatCaseService } from './format-case.service';
+import { SpacerComponent } from './spacer/spacer.component';
+import { KeysPipe } from './keys.pipe';
+import { CircularJsonPipe } from './circular-json.pipe';
 
 @NgModule({
     imports: [],
     exports: [
-        NumberSiPrefixPipe
+        NumberSiPrefixPipe,
+        KeysPipe,
+        CircularJsonPipe,
+        SpacerComponent
     ],
     declarations: [
-        NumberSiPrefixPipe
+        NumberSiPrefixPipe,
+        KeysPipe,
+        CircularJsonPipe,
+        SpacerComponent
     ],
     providers: [
         FormatCaseService